Layout & border & roundedCorners & etc.. use float instead of int...
Layout & border & roundedCorners & etc.. use float instead of int (WwBxA12K31)

+++ include/ssGUI/Extensions/AdvancedPosition.hpp
@@ -31,8 +31,8 @@ namespace ssGUI::Extensions
         bool HorizontalUsePercentage;
         bool VerticalUsePercentage;
 
-        int HorizontalPixelValue;
-        int VerticalPixelValue;
+        float HorizontalPixelValue;
+        float VerticalPixelValue;
 
         float HorizontalPercentageValue;
         float VerticalPercentageValue;
@@ -66,8 +66,8 @@ namespace ssGUI::Extensions
             bool HorizontalUsePercentage;
             bool VerticalUsePercentage;
 
-            int HorizontalPixelValue;
-            int VerticalPixelValue;
+            float HorizontalPixelValue;
+            float VerticalPixelValue;
 
             float HorizontalPercentageValue;
             float VerticalPercentageValue;
@@ -125,16 +125,16 @@ namespace ssGUI::Extensions
             virtual bool IsVerticalUsePercentage() const;
 
             //function: SetHorizontalPixel
-            virtual void SetHorizontalPixel(int pixel);
+            virtual void SetHorizontalPixel(float pixel);
 
             //function: GetHorizontalPixel            
-            virtual int GetHorizontalPixel() const;
+            virtual float GetHorizontalPixel() const;
 
             //function: SetVerticalPixel
-            virtual void SetVerticalPixel(int pixel);
+            virtual void SetVerticalPixel(float pixel);
 
             //function: GetVerticalPixel            
-            virtual int GetVerticalPixel() const;
+            virtual float GetVerticalPixel() const;
             
             //function:  SetHorizontalPercentage           
             //0 = 0%, 1 = 100%. This can go below 0 or over 1
